fix: only embed Azure roots on darwin
